+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.8.1] - 2026-03-26 [EXPORT-CONFIG]
 
 ### Added
+- Added import/export config.
 - Save SFTP presets globally.
 
 ### Changed